botlib: complete 217-function gcc272 ELF-oracle backlog sweep; fix FindField PE regression from project's first full-codebase audit
Consolidates a long multi-session effort that worked the entire SQL
`elf_candidates` backlog (217 functions flagged by
original 1999 gladi386.so) down to full disposition: 93 fixed to
byte-identical or improved ELF match, 62 investigated with a documented
reason the residual is a genuine cross-compiler floor, 28 accepted as
improved-but-not-byte-identical trade-offs, 20 excluded as out of scope,
14 independently reconfirmed. Every fix was PE-oracle (MSVC6 gladiator.dll)
regression-checked before being kept; full function-by-function rationale
elf_candidates table's per-row disposition notes.

This commit's own concrete contribution, on top of the accumulated
uncommitted work from prior sessions:

- Ran this project's first-ever whole-codebase PE regression audit
  (audit_dll.sh, 798 functions) instead of only the usual per-function
  spot check after each individual ELF fix. Diffed against the committed
  found 8 divergent rows. Root-caused all 8 using the project's
  git-stash-and-rebuild isolation methodology (stash all uncommitted
  work, rebuild the PE oracle from the last real commit, qaudit.sh the
  suspects, stash pop):

  - FindField (l_struct.c): a genuine, silent PE regression. An earlier
    session's uncommitted ELF-motivated change had re-applied the
    index-based loop idiom (`return &defs[i]; i += 7;`) that
    asm_matching_project.md's "FindField: a fourth confirmed
    irreconcilable conflict" section had *already* tried and rejected
    (and the SQL row already recorded as "PE-safe, left as OUR+4") --
    prior commits, undetected because no full sweep had been run since.
    Reverted to the documented pointer-with-redundant-`i` form

    stacked ELF fixes. Root cause: real's disasm uses genuine FPU
    all 3 vec3 components via `*(int*)&dst = *(int*)&src`. Changed just
    the Z-component write to a plain `bs->formationgoal.origin[2] =
    byte_diffs improved 186 -> 152, better than the original pre-sweep

  - BotUpdateEntityItems (be_ai_goal.c): PE byte_diffs grew 28 -> 199 as
    a side effect of an earlier, independently-correct struct-by-value
    retype fix (float entinfo[31] -> aas_entityinfo_t entinfo, matching
    the established AINode_Battle_Fight pattern). Investigated; traced to
    allocation-path block. Never crossed a MATCH-category boundary
    (INSN_COUNT_MATCH both before and after) -- accepted as a documented
    trade-off, not reverted.

    from MATCH to INSN_COUNT_MATCH, but neither function's file
    (be_aas_sample.c, be_ai_chat.c) was touched by any of this sweep's
    commits. Both are part of this project's previously-documented
    established earlier as non-deterministic by rebuilding an unmodified
    baseline twice and finding disagreement between the two builds. This
    session re-ran that exact test for these 2 functions specifically
    (a second, genuinely independent rebuild via FORCE=1, which bypasses
    the build script's own mtime cache -- a plain second invocation just
    reuses the first build's bytes and is not a real second data point).
    Result: both independent builds of byte-identical committed source
    produced the SAME INSN_COUNT_MATCH value for both functions -- the
    opposite of the earlier AnglesToAxis finding. This reclassifies these
    2 rows from "randomly flaky" to "reproducibly drifted": the
    dll_audit.tsv was last generated, in a way that now consistently
    reproduces a worse layout for these two FP-dot-product-order-sensitive
    functions. No source change made for either.

  - AAS_TraceBSPModel, sub_100056D0 (be_aas_bspq2.c), AAS_Reachability_Ladder:
    confirmed as unexplained-but-welcome positive ripple from earlier
    byte-diff shifts. No action needed.

- Refreshed both evidence files from a from-scratch full sweep of the
  fully-fixed tree, diffed address-by-address against their previous
  committed versions to confirm only genuine, explained changes were
  recorded:
      (810 rows). Zero regressions; 27 rows newly reached MATCH; 30 more
      improved byte_diffs within their existing status bucket.
      rows) -- the net -2 is entirely and exactly the two reproducibly-
      drifted rows above; excluding those, zero true PE regressions from
      any ELF-side fix in the whole 217-function sweep, and 3 rows
      improved (AAS_TraceBSPModel, sub_100056D0, BotGetFormationGoal).

- Verified via a second full audit_dll.sh sweep after the fixes (zero
  MISSING, zero unexplained rows), a native `make clean && make botlib
  game` build (Linux aarch64, clean except pre-existing warnings), and a
  full engine smoke test (quake2 +set game gladiator +set deathmatch 1
  +exec autoexec.cfg +map q2dm1): reached bot spawn (`loaded 18 bots`,
  `AAS initialized.`, `Laura Craft entered the game`) and a clean shutdown
  don't break runtime behavior.

(new "Final full-codebase PE regression audit" capstone section, plus an
addendum on the pre-existing FindField section) and updated MEMORY.md's
index line. Established as a standing project requirement: run a full
audit_dll.sh + audit_so.py pair and refresh both evidence files before any
session that intends to commit a batch of ELF backlog work, not only when
a specific function is suspected -- this is the only mechanism that has
ever caught a silent multi-session regression (FindField) or reliably
distinguished stale evidence from a real regression.

Source-commit: 7bc61e7138f0b035281fd001dfba385bbd8339ea

+++ botlib/be_ai2_main.c
@@ -222,7 +222,7 @@ int __cdecl BotChangeViewAngles(bot_state_t *bs, float thinktime)
   {
     /* AngleDifference's FPU return feeds straight into the int conversion and
      * abs: v5 = abs((int)AngleDifference(viewangles[i], ideal_viewangles[i])). */
-    v5 = (float)abs((__int64)AngleDifference(*(v3 - 6), *(v3 - 3)));
+    v5 = (float)abs((int)AngleDifference(*(v3 - 6), *(v3 - 3)));
     if ( v5 > *v3 )
     {
       v6 = v10 + *v3;
